Refactored layer address usage, using type definition for L2 (MAC) and L3 (Network) addresses and switched filterSignal parameter from Signal& to AirFrame*.
Prepared the INET MACAddress usage for L2-Addresses and uniform type definition for L3 addresses, which was switching from int, long and other definitions. The SimpleAddress.h introduced a new class for the layer addresses, the old enum was removed.

+++ examples/analogueModels/SimpleMacLayer.h
@@ -25,13 +25,13 @@ protected:
 	MacToPhyInterface* phy;
 
 	/** @brief Omnet gates.*/
-	int dataOut;
-	int dataIn;
+	int  dataOut;
+	int  dataIn;
 
-	int myIndex;
+	long myIndex;
 
 	/** @brief The index of the host the next packet should be send to. */
-	int nextReceiver;
+	long nextReceiver;
 
 	enum {
 		/** @brief Used as Message kind to identify MacPackets. */
